ml/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-multipart
transformers
accelerate
